requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.0
langchain==0.0.300
pinecone-client==2.2.4
python-multipart==0.0.6
pandas==1.5.3
numpy==1.24.0
tiktoken==0.3.0